+++ src/containers/AddFunds/CoinSelectorModal/CoinCard/CoinCard.types.ts
@@ -1,0 +1,6 @@
+import { ICoin } from '@helpers/coins';
+
+export interface CoinCardProps {
+	coin: ICoin;
+	onSelect: Function;
+}
